[json] Removed paremus code
